api_logic/requirements.txt
backcall==0.2.0
beautifulsoup4==4.9.3
bert-extractive-summarizer==0.7.1
blis==0.7.4
boto==2.49.0
boto3 @ file:///tmp/build/80754af9/boto3_1624558601899/work
botocore @ file:///tmp/build/80754af9/botocore_1624291157435/work
brotlipy==0.7.0
bz2file==0.98
catalogue==2.0.4
certifi==2021.5.30
cffi @ file:///tmp/build/80754af9/cffi_1613246951236/work
chardet @ file:///tmp/build/80754af9/chardet_1607706775000/work
click==7.1.2
cryptography @ file:///tmp/build/80754af9/cryptography_1616767007030/work
cssselect==1.1.0
cymem==2.0.5
Cython @ file:///tmp/build/80754af9/cython_1618435158654/work
decorator==5.0.9
filelock==3.0.12
gensim @ file:///tmp/build/80754af9/gensim_1621558514262/work
huggingface-hub==0.0.12
idna @ file:///home/linux1/recipes/ci/idna_1610986105248/work
ipython==7.25.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==3.0.1
jmespath @ file:///tmp/build/80754af9/jmespath_1594304593830/work
joblib==1.0.1
lxml==4.6.3
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mkl-fft==1.3.0
mkl-random @ file:///tmp/build/80754af9/mkl_random_1618853971975/work
mkl-service==2.3.0
murmurhash==1.0.5
nltk==3.6.2
numpy==1.21.0
olefile==0.46
packaging==20.9
parso==0.8.2
pathy==0.6.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow @ file:///tmp/build/80754af9/pillow_1617385572585/work
preshed==3.0.5
prompt-toolkit==3.0.19
ptyprocess==0.7.0
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pydantic==1.7.4
Pygments==2.9.0
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1608057966937/work
pyparsing==2.4.7
pyquery==1.4.3
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
pysummarization==1.1.7
python-dateutil @ file:///home/ktietz/src/ci/python-dateutil_1611928101742/work
python-readability==0.0.0
PyYAML==5.4.1
readability-lxml==0.8.1
regex==2021.4.4
requests @ file:///tmp/build/80754af9/requests_1608241421344/work
s3transfer @ file:///tmp/build/80754af9/s3transfer_1623278116614/work
sacremoses==0.0.45
scikit-learn==0.24.2
scipy==1.7.0
sentencepiece==0.1.96
six @ file:///tmp/build/80754af9/six_1623709665295/work
smart-open==5.1.0
soupsieve==2.2.1
spacy==3.0.6
spacy-legacy==3.0.6
srsly==2.4.1
thinc==8.0.6
threadpoolctl==2.1.0
tokenizers==0.10.3
torch==1.9.0
torchvision==0.2.2
tqdm==4.61.1
traitlets==5.0.5
transformers==4.8.1
typer==0.3.2
typing-extensions==3.10.0.0
urllib3==1.26.6
wasabi==0.8.2
wcwidth==0.2.5
